testapi/requirements.txt
openai~=0.27.4
django~=5.0.4
djangorestframework~=3.15.1
requests~=2.31.0
pymysql~=1.1.0